refactor: move april tag field layout into Field.java + small cleanup

+++ src/main/java/com/stuypulse/robot/subsystems/vision/VisionIO.java
@@ -56,7 +56,7 @@ public interface VisionIO {
 
     public int pipeline = 0;
 
-    public double[] aprilTagIDWhitelist = Field.ALL_TAGS;
+    public int[] aprilTagIDWhitelist = Field.ALL_TAGS;
 
     public int imuMode = 1;
   }
